This commit was generated by cvs2svn to compensate for changes in r5824, which included commits to RCS files with non-trunk default branches.
svn path=/trunk/; revision=5825

+++ applications/extractor/extract/src/interpolate.c
@@ -9,7 +9,7 @@
 *
 *	Contents:	Interpolation of input data.
 *
-*	Last modify:	29/04/98
+*	Last modify:	05/04/99
 *
 *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 */
@@ -34,7 +34,7 @@ void    init_interpolate(picstruct *field, int xtimeout, int ytimeout)
   if (ytimeout>=0)
     {
     QMALLOC(field->interp_ytimeoutbuf, int, field->width);
-    memset(field->interp_ytimeoutbuf, 0, field->width*sizeof(int));
+    memset(field->interp_ytimeoutbuf, 0, (size_t)(field->width*sizeof(int)));
     }
 
   field->interp_xtimeout = xtimeout;
